Extracted class declarations out of ojph_codestream_local.h
