+++ src/core/codestream/ojph_codestream_local.h
@@ -42,6 +42,12 @@
 #include "ojph_defs.h"
 #include "ojph_file.h"
 #include "ojph_params_local.h"
+#include "ojph_codeblock.h"
+#include "ojph_subband.h"
+#include "ojph_precinct.h"
+#include "ojph_resolution.h"
+#include "ojph_tile_comp.h"
+#include "ojph_tile.h"
 
 namespace ojph {
 
@@ -163,289 +169,8 @@ namespace ojph {
       outfile_base *outfile;
       infile_base *infile;
     };
+  
 
-    //////////////////////////////////////////////////////////////////////////
-    class tile
-    {
-    public:
-      static void pre_alloc(codestream *codestream, const rect& tile_rect,
-                            const rect& recon_tile_rect);
-      void finalize_alloc(codestream *codestream, const rect& tile_rect,
-                          const rect& recon_tile_rect, 
-                          ui32 tile_idx, ui32 offset);
-
-      bool push(line_buf *line, ui32 comp_num);
-      void prepare_for_flush();
-      void fill_tlm(param_tlm* tlm);
-      void flush(outfile_base *file);
-      void parse_tile_header(const param_sot& sot, infile_base *file,
-                             const ui64& tile_start_location);
-      bool pull(line_buf *, ui32 comp_num);
-      rect get_tile_rect() { return tile_rect; }
-
-    private:
-      //codestream *parent;
-      rect tile_rect, recon_tile_rect;
-      ui32 num_comps;
-      tile_comp *comps;
-      ui32 num_lines;
-      line_buf* lines;
-      bool reversible, employ_color_transform, resilient;
-      rect *comp_rects, *recon_comp_rects;
-      ui32 *line_offsets;
-      ui32 skipped_res_for_read;
-
-      ui32 *num_bits;
-      bool *is_signed;
-      ui32 *cur_line;
-      int prog_order;
-
-    private:
-      param_sot sot;
-      int next_tile_part;
-
-    private:
-      int profile;
-      ui32 *num_comp_bytes; //this for use with TLM
-    };
-
-    //////////////////////////////////////////////////////////////////////////
-    class tile_comp
-    {
-    public:
-      static void pre_alloc(codestream *codestream, const rect& comp_rect,
-                            const rect& recon_comp_rect);
-      void finalize_alloc(codestream *codestream, tile *parent,
-                          ui32 comp_num, const rect& comp_rect,
-                          const rect& recon_comp_rect);
-
-      ui32 get_num_resolutions() { return num_decomps + 1; }
-      ui32 get_num_decompositions() { return num_decomps; }
-      tile* get_tile() { return parent_tile; }
-      line_buf* get_line();
-      void push_line();
-      line_buf* pull_line();
-
-      ui32 prepare_precincts();
-      void write_precincts(ui32 res_num, outfile_base *file);
-      bool get_top_left_precinct(ui32 res_num, point &top_left);
-      void write_one_precinct(ui32 res_num, outfile_base *file);
-      void parse_precincts(ui32 res_num, ui32& data_left, infile_base *file);
-      void parse_one_precinct(ui32 res_num, ui32& data_left, 
-                              infile_base *file);
-
-    private:
-      tile *parent_tile;
-      resolution *res;
-      rect comp_rect;
-      ojph::point comp_downsamp;
-      ui32 num_decomps;
-      ui32 comp_num;
-    };
-
-    //////////////////////////////////////////////////////////////////////////
-    class resolution
-    {
-    public:
-
-    public:
-      static void pre_alloc(codestream *codestream, const rect& res_rect,
-                            const rect& recon_res_rect, ui32 res_num);
-      void finalize_alloc(codestream *codestream, const rect& res_rect,
-                          const rect& recon_res_rect, ui32 comp_num,
-                          ui32 res_num, point comp_downsamp,
-                          tile_comp *parent_tile_comp,
-                          resolution *parent_res);
-
-      line_buf* get_line() { return lines + 0; }
-      void push_line();
-      line_buf* pull_line();
-      rect get_rect() { return res_rect; }
-      ui32 get_comp_num() { return comp_num; }
-
-      ui32 prepare_precinct();
-      void write_precincts(outfile_base *file);
-      bool get_top_left_precinct(point &top_left);
-      void write_one_precinct(outfile_base *file);
-      resolution *next_resolution() { return child_res; }
-      void parse_all_precincts(ui32& data_left, infile_base *file);
-      void parse_one_precinct(ui32& data_left, infile_base *file);
-
-    private:
-      bool reversible, skipped_res_for_read, skipped_res_for_recon;
-      ui32 num_lines;
-      ui32 num_bands, res_num;
-      ui32 comp_num;
-      point comp_downsamp;
-      rect res_rect;
-      line_buf *lines;
-      subband *bands;
-      tile_comp *parent_comp;
-      resolution *parent_res, *child_res;
-      //precincts stuff
-      precinct *precincts;
-      size num_precincts;
-      size log_PP;
-      ui32 max_num_levels;
-      int tag_tree_size;
-      ui32 level_index[20]; //more than enough
-      point cur_precinct_loc; //used for progressing spatial modes (2, 3, 4)
-      //wavelet machinery
-      ui32 cur_line;
-      bool vert_even, horz_even;
-      mem_elastic_allocator *elastic;
-    };
-
-    //////////////////////////////////////////////////////////////////////////
-    struct precinct
-    {
-      precinct() {
-        scratch = NULL; bands = NULL; coded = NULL;
-        num_bands = 0; may_use_sop = uses_eph = false;
-      }
-      ui32 prepare_precinct(int tag_tree_size, ui32* lev_idx,
-                            mem_elastic_allocator *elastic);
-      void write(outfile_base *file);
-      void parse(int tag_tree_size, ui32* lev_idx,
-                 mem_elastic_allocator *elastic,
-                 ui32& data_left, infile_base *file, bool skipped);
-
-      ui8 *scratch;
-      point img_point;   //the precinct projected to full resolution
-      rect cb_idxs[4]; //indices of codeblocks
-      subband *bands;  //the subbands
-      coded_lists* coded;
-      ui32 num_bands;
-      bool may_use_sop, uses_eph;
-    };
-
-    //////////////////////////////////////////////////////////////////////////
-    class subband
-    {
-      friend struct precinct;
-    public:
-      static void pre_alloc(codestream *codestream, const rect& band_rect,
-                            ui32 res_num);
-      void finalize_alloc(codestream *codestream, const rect& band_rect,
-                          resolution* res, ui32 res_num, ui32 subband_num);
-
-      void exchange_buf(line_buf* l);
-      line_buf* get_line() { return lines; }
-      void push_line();
-
-      void get_cb_indices(const size& num_precincts, precinct *precincts);
-      float get_delta() { return delta; }
-
-      line_buf* pull_line();
-
-    private:
-      ui32 res_num, band_num;
-      bool reversible;
-      bool empty;
-      rect band_rect;
-      line_buf *lines;
-      resolution* parent;
-      codeblock* blocks;
-      size num_blocks;
-      size log_PP;
-      ui32 xcb_prime, ycb_prime;
-      ui32 cur_cb_row;
-      int cur_line;
-      int cur_cb_height;
-      float delta, delta_inv;
-      ui32 K_max;
-      coded_cb_header *coded_cbs;
-      mem_elastic_allocator *elastic;
-    };
-
-    //////////////////////////////////////////////////////////////////////////
-    class codeblock
-    {
-      friend struct precinct;
-    public:
-      static void pre_alloc(codestream *codestream, const size& nominal);
-      void finalize_alloc(codestream *codestream, subband* parent,
-                          const size& nominal, const size& cb_size,
-                          coded_cb_header* coded_cb,
-                          ui32 K_max, int tbx0);
-      void push(line_buf *line);
-      void encode(mem_elastic_allocator *elastic);
-      void recreate(const size& cb_size, coded_cb_header* coded_cb);
-
-      void decode();
-      void pull_line(line_buf *line);
-
-    private:
-      ui32* buf;
-      size nominal_size;
-      size cb_size;
-      ui32 stride;
-      ui32 buf_size;
-      subband* parent;
-      int line_offset;
-      ui32 cur_line;
-      float delta, delta_inv;
-      ui32 K_max;
-      bool reversible;
-      bool resilient;
-      bool stripe_causal;
-      bool zero_block; // true when the decoded block is all zero
-      ui32 max_val[8]; // supports up to 256 bits
-      coded_cb_header* coded_cb;
-
-    private:
-      // define function signature simple memory clearing
-      typedef void (*mem_clear_fun)(void* addr, size_t count);
-      // a pointer to the max value finding function
-      mem_clear_fun mem_clear;
-      static void gen_mem_clear(void* addr, size_t count);
-
-      // define function signature for max value finding
-      typedef ui32 (*find_max_val_fun)(ui32* addr);
-      // a pointer to the max value finding function
-      find_max_val_fun find_max_val;
-      static ui32 gen_find_max_val(ui32* addr) { return addr[0]; }
-
-      // define line transfer function signature from subbands to codeblocks
-      typedef void (*tx_to_cb_fun)(const void *sp, ui32 *dp, ui32 K_max,
-                                   float delta_inv, ui32 count, ui32* max_val);
-      // a pointer to function transferring samples from subbands to codeblocks
-      tx_to_cb_fun tx_to_cb;
-      static void gen_rev_tx_to_cb(const void *sp, ui32 *dp, ui32 K_max,
-                                   float delta_inv, ui32 count, ui32* max_val);
-      static void gen_irv_tx_to_cb(const void *sp, ui32 *dp, ui32 K_max,
-                                   float delta_inv, ui32 count, ui32* max_val);
-
-      // define line transfer function signature from codeblock to subband
-      typedef void (*tx_from_cb_fun)(const ui32 *sp, void *dp, ui32 K_max,
-                                     float delta, ui32 count);
-      // a pointer to function transferring samples from codeblocks to subbands
-      tx_from_cb_fun tx_from_cb;
-      static void gen_rev_tx_from_cb(const ui32 *sp, void *dp, ui32 K_max,
-                                     float delta, ui32 count);
-      static void gen_irv_tx_from_cb(const ui32 *sp, void *dp, ui32 K_max,
-                                     float delta, ui32 count);
-
-      // define the block decoder function signature
-      typedef bool (*cb_decoder_fun)(ui8* coded_data, ui32* decoded_data,
-        ui32 missing_msbs, ui32 num_passes, ui32 lengths1, ui32 lengths2,
-        ui32 width, ui32 height, ui32 stride, bool stripe_causal); 
-      // a pointer to the decoder function
-      static cb_decoder_fun decode_cb;
-    };
-
-    //////////////////////////////////////////////////////////////////////////
-    struct coded_cb_header
-    {
-      ui32 pass_length[2];
-      ui32 num_passes;
-      ui32 Kmax;
-      ui32 missing_msbs;
-      coded_lists *next_coded;
-
-      static const int prefix_buf_size = 8;
-      static const int suffix_buf_size = 16;
-    };
 
     //////////////////////////////////////////////////////////////////////////
     void sse_mem_clear(void* addr, size_t count);
@@ -487,6 +212,5 @@ namespace ojph {
 
   }
 }
-
 
 #endif // !OJPH_CODESTREAM_LOCAL_H
